feat: Update CSRF_TRUSTED_ORIGINS and CSRF_COOKIE_SECURE in azure.py
The recent code changes modify the `CSRF_TRUSTED_ORIGINS` and `CSRF_COOKIE_SECURE` variables in `azure.py`. The updated values for `CSRF_TRUSTED_ORIGINS` are `'apprefugiodeperros.azurewebsites.net'` and `CSRF_COOKIE_SECURE` is set to `False`.

Based on the recent user commits and repository commits, it seems that these changes are related to updating forms and views in the project.

Please review and consider this commit message: